+++ rules/node.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = {
+  rules: {
+    'callback-return': 'error',
+    'handle-callback-err': 'error',
+    'no-mixed-requires': 'error'
+  }
+};