+++ rules/variables.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  rules: {
+    'no-undefined': 'error',
+    'no-use-before-define': 'error'
+  }
+};